src/entities/Profile/model/services/validateProfileData/validateProfileData.test.ts
import { Currency } from 'entities/Currency';
import { Country } from 'entities/Country';

import { validateProfileData } from './validateProfileData';

import { ValidateProfileError } from '../../types/profile';

const data = {
  username: 'admin',
  first: 'dima',
  lastname: 'tester',
  age: 23,
  city: 'New York',
  currency: Currency.RUB,
  country: Country.BELARUS,
  avatar: 'https://gorodprizrak.com/wp-content/uploads/2023/03/13b5a372655fc489f4a29217d43c235d.jpg',
};

describe('validateProfileData', () => {
  test('should return no errors', () => {
    const result = validateProfileData(data);

    expect(result).toEqual([]);
  });

  test('should return user data error', () => {
    const result = validateProfileData({ ...data, username: '' });

    expect(result).toEqual([ValidateProfileError.INCORRECT_USER_DATA]);
  });

  test('should return empty data error', () => {
    const result = validateProfileData();

    expect(result).toEqual([ValidateProfileError.NO_DATA]);
  });

  test('should return empty data error', () => {
    const result = validateProfileData({});

    expect(result).toEqual([
      ValidateProfileError.INCORRECT_USER_DATA,
      ValidateProfileError.INCORRECT_AGE,
      ValidateProfileError.INCORRECT_COUNTRY,
    ]);
  });

  test('should return age error', () => {
    const result = validateProfileData({ ...data, age: undefined });

    expect(result).toEqual([ValidateProfileError.INCORRECT_AGE]);
  });

  test('should return country error', () => {
    const result = validateProfileData({ ...data, country: undefined });

    expect(result).toEqual([ValidateProfileError.INCORRECT_COUNTRY]);
  });
});
